+++ Locales/enUS.lua
@@ -68,6 +68,12 @@ select(2, ...).L = setmetatable({
     ["SNIPPETS_TIPS"] = "[Double-Click] to rename. [Shift-Click] to delete. All checked snippets will be automatically invoked at the end of Cell initialization process (in ADDON_LOADED event).",
 
     ["CHANGELOGS"] = [[
+        <h1>r162-release (Apr 14, 2023, 19:00 GMT+8)</h1>
+        <p>* Moved "Sort By Role" to Layouts.</p>
+        <p>* Added "Hide Self" in Layouts.</p>
+        <p>* Fixed "Track by name" for custom indicators (wrath).</p>
+        <br/>
+
         <h1>r161-release (Apr 8, 2023, 20:00 GMT+8)</h1>
         <p>* Fixed dispellable debuff type checker (retail).</p>
         <p>* Updated zhTW.</p>