+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r160-release...ea10b9f776ada3b6fdd348f8e9b8ee40d05f4fed)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r161-release...b6c08cae2b1c4c5e724567bad7b9b96799249149)
 
-- Fix dispelling checker (retail)
-- Update zhTW
+- Fix trackByName for wrath indicators
+- Add hideSelf for party
+- Move "Sort By Role" to Layouts